Update deleting an image from the storage

+++ components/Post.jsx
@@ -55,7 +55,10 @@ export default function Post({ post }) {
   async function deletePost() {
     if (window.confirm("Are you sure you want to delete this post?")) {
       deleteDoc(doc(db, "posts", post.id)); //delete from database
-      deleteObject(ref(storage, `posts/${post.id}/image`)); //delete from storage
+      if(post.data().image){
+        deleteObject(ref(storage, `posts/${post.id}/image`)); //delete from storage
+      }
+      
     }
   }
